syslog is not available in linux (or plan9, but meh)
